requirements.txt
djangorestframework
python-decouple
Pillow
psycopg2-binary
django-rest-auth 
django-allauth
